Moved `FrameworkBuilder` to its own file
